refactor(import): lazy-load implementations of import sources
