+++ src/livecodes/import/code.ts
@@ -2,8 +2,6 @@ import type { Config } from '../models';
 // eslint-disable-next-line import/no-internal-modules
 import { decompress } from '../utils/compression';
 
-export const isCompressedCode = (url: string) => url.startsWith('code/');
-
 export const importCompressedCode = (url: string) => {
   const code = url.slice(5);
   let config: Partial<Config>;
