+++ src/livecodes/import/project-id.ts
@@ -1,7 +1,5 @@
 import { shareService } from '../services';
 
-export const isProjectId = (url: string) => url.startsWith('id/');
-
 export const importProject = (url: string) => {
   const id = url.slice(3);
   return shareService.getProject(id);
